+++ src/renderingTools.ts
@@ -8,8 +8,17 @@ import {
 import { renderSingleLineBoxV1 } from "./renderLineToolV1";
 import { renderSingleLineBoxV2 } from "./renderLineToolV2";
 import { renderSingleLineBoxV3 } from "./renderLineToolV3";
+import { renderSingleLineBoxV4 } from "./renderLineToolV4";
+import {
+    AdvancedColoringFields,
+    editorBackgroundFormula,
+    makeInnerKitchenNotation,
+} from "./settingsManager";
 import { IBlockRender } from "./utils";
 import { notYetDisposedDecsObject } from "./utils2";
+
+const neu = "neutral";
+const advNeu = makeInnerKitchenNotation("basic");
 
 export interface ISingleLineBox {
     editorInfo: IEditorInfo;
@@ -42,6 +51,9 @@ export interface ISingleLineBox {
         lineZero: number;
         inLineIndexZero: number;
     };
+    inputBorderColor: string;
+    inputBackgroundColor: string;
+    borderSize: number;
 }
 
 export const renderSingleBlock = ({
@@ -63,6 +75,206 @@ export const renderSingleBlock = ({
     if (!firstVisibleChar || firstVisibleChar.lineZero < 0) {
         return;
     }
+
+    // -----11111111111
+    let inputBorderColor: string = `linear-gradient(to right, ${"transparent"}, ${"transparent"})`; // in final state it always must be linear gradient
+    let inputBackgroundColor: string = `linear-gradient(to right, ${editorBackgroundFormula}, ${editorBackgroundFormula})`; // in final state it always must be linear gradient
+
+    inputBorderColor = glo.coloring.border;
+
+    switch (depth) {
+        case 0:
+            inputBackgroundColor = glo.coloring.onEachDepth[0];
+            inputBorderColor = glo.coloring.borderOfDepth0;
+            break;
+
+        case 1:
+            inputBackgroundColor = glo.coloring.onEachDepth[1];
+            break;
+        case 2:
+            inputBackgroundColor = glo.coloring.onEachDepth[2];
+            break;
+        case 3:
+            inputBackgroundColor = glo.coloring.onEachDepth[3];
+            break;
+        case 4:
+            inputBackgroundColor = glo.coloring.onEachDepth[4];
+            break;
+        case 5:
+            inputBackgroundColor = glo.coloring.onEachDepth[5];
+            break;
+        case 6:
+            inputBackgroundColor = glo.coloring.onEachDepth[6];
+            break;
+        case 7:
+            inputBackgroundColor = glo.coloring.onEachDepth[7];
+            break;
+        case 8:
+            inputBackgroundColor = glo.coloring.onEachDepth[8];
+            break;
+        case 9:
+            inputBackgroundColor = glo.coloring.onEachDepth[9];
+            break;
+        case 10:
+            inputBackgroundColor = glo.coloring.onEachDepth[10];
+            break;
+
+        default:
+            inputBackgroundColor = glo.coloring.onEachDepth[10];
+    }
+
+    let borderSize = glo.borderSize;
+    if (glo.enableFocus && isFocusedBlock) {
+        if (!glo.coloring.focusedBlock.includes("same")) {
+            inputBackgroundColor = glo.coloring.focusedBlock;
+        }
+        if (!glo.coloring.borderOfFocusedBlock.includes("same")) {
+            inputBorderColor = glo.coloring.borderOfFocusedBlock;
+        }
+
+        borderSize = 2;
+    }
+
+    // kkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk
+
+    let borderOfAC: undefined | string = undefined;
+    let backgroundOfAC: undefined | string = undefined;
+
+    const aCSettingsForBorders = glo.coloring.advanced.forBorders;
+    const aCSettingsForBackgrounds = glo.coloring.advanced.forBackgrounds;
+
+    const focPath = editorInfo.focusTreePath;
+    // if (focPath) {
+    //     if (focPath[depth] === inDepthBlockIndex) {
+    //         // borderOfAC =
+    //         //     "linear-gradient(to right, red, rgba(250, 10, 100, 0.3))";
+    //     }
+    // }
+
+    const focInners = editorInfo.innersFromFocus;
+    // if (focInners) {
+    //     const indexes = focInners[depth];
+    //     if (indexes && indexes.includes(inDepthBlockIndex)) {
+    //         // borderOfAC = "linear-gradient(to right, blue, blue)";
+    //     }
+    // }
+
+    const focusedBlockInfo = editorInfo.focusDuo.curr;
+    const focDepth = focusedBlockInfo?.depth;
+
+    // const focIndITD = focusedBlockInfo.indexInTheDepth;
+
+    const inOnFocTree =
+        focDepth &&
+        depth <= focDepth &&
+        focPath &&
+        focPath[depth] === inDepthBlockIndex;
+
+    const isInsideFoc =
+        focDepth &&
+        depth >= focDepth &&
+        focInners &&
+        focInners[depth]?.includes(inDepthBlockIndex);
+
+    const decideColor = (
+        aCSettings: {
+            priority: number;
+            sequence: string[];
+            kind: AdvancedColoringFields;
+        }[],
+    ): string | undefined => {
+        for (let i = 0; i < aCSettings.length; i += 1) {
+            const seqInfo = aCSettings[i];
+            const kind = seqInfo.kind;
+
+            if (kind === AdvancedColoringFields.fromD0ToInward_All) {
+                const myColor = seqInfo.sequence[depth];
+                if (myColor && myColor !== neu) {
+                    // borderOfAC = myColor;
+                    // break;
+                    if (myColor === advNeu) {
+                        return undefined;
+                    }
+                    return myColor;
+                }
+            } else if (
+                kind === AdvancedColoringFields.fromD0ToInward_FocusTree
+            ) {
+                let myColor = neu;
+                if (inOnFocTree) {
+                    myColor = seqInfo.sequence[depth];
+                }
+                if (myColor && myColor !== neu) {
+                    // borderOfAC = myColor;
+                    // break;
+                    if (myColor === advNeu) {
+                        return undefined;
+                    }
+                    return myColor;
+                }
+            } else if (kind === AdvancedColoringFields.fromFocusToInward_All) {
+                let myColor = neu;
+                if (focDepth && isInsideFoc) {
+                    const myInd = depth - focDepth;
+                    myColor = seqInfo.sequence[myInd];
+                }
+                if (myColor && myColor !== neu) {
+                    // borderOfAC = myColor;
+                    // break;
+                    if (myColor === advNeu) {
+                        return undefined;
+                    }
+                    return myColor;
+                }
+            } else if (kind === AdvancedColoringFields.fromFocusToOutward_All) {
+                let myColor = neu;
+                if (focDepth && depth <= focDepth) {
+                    const myInd = focDepth - depth;
+                    myColor = seqInfo.sequence[myInd];
+                }
+                if (myColor && myColor !== neu) {
+                    // borderOfAC = myColor;
+                    // break;
+                    if (myColor === advNeu) {
+                        return undefined;
+                    }
+                    return myColor;
+                }
+            } else if (
+                kind === AdvancedColoringFields.fromFocusToOutward_FocusTree
+            ) {
+                let myColor = neu;
+                if (focDepth && inOnFocTree) {
+                    const myInd = focDepth - depth;
+                    myColor = seqInfo.sequence[myInd];
+                }
+                if (myColor && myColor !== neu) {
+                    // borderOfAC = myColor;
+                    // break;
+                    if (myColor === advNeu) {
+                        return undefined;
+                    }
+                    return myColor;
+                }
+            }
+        }
+
+        return undefined;
+    };
+
+    borderOfAC = decideColor(aCSettingsForBorders);
+    backgroundOfAC = decideColor(aCSettingsForBackgrounds);
+
+    // ------222222222
+
+    if (borderOfAC) {
+        inputBorderColor = borderOfAC;
+    }
+    if (backgroundOfAC) {
+        inputBackgroundColor = backgroundOfAC;
+    }
+
+    // ------- 333333
 
     for (
         let currLineZero = firstVisibleChar.lineZero;
@@ -165,6 +377,9 @@ export const renderSingleBlock = ({
 
             firstVisibleChar,
             lastVisibleChar,
+            inputBorderColor,
+            inputBackgroundColor,
+            borderSize,
         };
 
         // renderSingleLineBoxV1(singleRangeRendArg); // old renderer function
@@ -178,7 +393,8 @@ export const renderSingleBlock = ({
             currLineZero >= firstLineOfMiddles &&
             currLineZero <= lastLineOfMiddles;
 
-        renderSingleLineBoxV3(singleRangeRendArg); // super new renderer function
+        // renderSingleLineBoxV3(singleRangeRendArg);
+        renderSingleLineBoxV4(singleRangeRendArg);
         if (isMid) {
             currLineZero = lastLineOfMiddles;
         }